Remove excerpts from comprehensive rules
I had originally thought that citing the comprehensive rules from
within the implementation would be good documentation and improve the
odds of implementing the rules correctly.

Sadly, the comprehensive rules were not written with an automated
implementation in mind, and as such, I feel that my attempt to cite
them all over the place is doing more harm than good.